requirements.txt
pandas
numpy
scikit-learn
umap-learn
matplotlib
seaborn
streamlit
